CHANGE(rawx): Allow to set internal buffer sizes
 ##### SUMMARY

Depending on the customer's workload, it may be useful to change the
internal buffer sizes. Benchmarks show that large buffer sizes are good
for big uploads, but may slow down small uploads.

As of 19.04 version, the internal chunk buffer is 2048 kiB long. For
workloads creating many chunks smaller than 1 MiB, it may be a good idea
to reduce it.

 ##### IMPACT
N/A

 ##### ADDITIONAL INFORMATION